Refactoring space, cleaning everithing
